+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - [feat (#21): [syntax/scanner, syntax/parser, typechecker, codegen] implement struct types & constructors](https://github.com/amp-lang/amp/issues/21)
 - [feat (#23): [cli] implement multiple module linking](https://github.com/amp-lang/amp/issues/23)
 - [feat (#22): [syntax/scanner, syntax/parser, typechecker, codegen] implement struct field accessing](https://github.com/amp-lang/amp/issues/22)
+- [feat (#24): [typechecker, codegen] assign fields of struct references](https://github.com/amp-lang/amp/issues/24)
+- [feat (#25): [typechecker, codegen] reference a struct field](https://github.com/amp-lang/amp/issues/25)
 
 ## v0.2.0-alpha